backend/pyproject.toml
[project]
name = "semantic-search-newsgroups"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.2",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.1",
    "huggingface-hub>=1.6.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "scikit-fuzzy>=0.5.0",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.2.3",
    "tqdm>=4.67.3",
    "umap-learn>=0.5.11",
]
